Initial work on handling network announcements
Right now, the app parses the news RSS feed from the official website
directly. The intention is for the server to begin acting as a
middle-man, so we can also read posts from their Facebook page
without leaking our Facebook API access tokens. Furthermore, the
server will be able to continuously check for updates and issue
notifications to clients whenever a new announcement is posted.