Begin work on MQTT gateway
"Gateways" are the new form of real-time communication between the
UnderLX server and clients. They are used to support communication
paradigms beyond what can be conveniently supported by the REST API.

Clients use the REST API to discover available gateways and optionally
connect to the ones they support.

This first gateway will use MQTT as its protocol, with the server acting
as the MQTT broker. Clients authenticate using the same credentials that
are used with the REST API. The broker controls message propagation to
ensure that messages published to arbitrary MQTT topics by clients don't
reach other clients.

This begins paving the way to the next generation of real-time features
in UnderLX, which can feature, among other things, the broadcast of
estimated vehicle arrival times to clients.